# Performance improvement kadi states