@jeanconn 2023-11-20
With https://github.com/sot/cxotime/pull/42 and https://github.com/sot/chandra_maneuver/pull/26 tests are now passing in for this PR (and master).